+++ clown_sort/sort_selector.py
@@ -11,7 +11,7 @@ from rich.text import Text
 
 from clown_sort.config import Config
 from clown_sort.filename_extractor import FilenameExtractor
-from clown_sort.files.image_file import ImageFile
+#from clown_sort.files.image_file import ImageFile
 from clown_sort.util.dict_helper import get_dict_key_by_value
 from clown_sort.util.logging import log
 from clown_sort.util.rich_helper import bullet_text, console, indented_bullet
@@ -24,7 +24,7 @@ SKIP = 'Skip'
 EXIT = 'Exit'
 
 
-def process_file(image: ImageFile) -> None:
+def process_file_with_popup(image: 'ImageFile') -> None:
     import PySimpleGUI as sg
     console.print(Panel(Text("Processing: '").append(str(image.file_path), style='cyan reverse').append("'..."), expand=False))
     extracted_text = ' '.join((image.extracted_text() or '').splitlines())